Implement AI Events dashboard with filtering, sorting, and media rendering; update violations schema to include reviewedAt timestamp
