+++ src/app/(pages)/layout.tsx
@@ -1,10 +1,13 @@
 'use client';
 import DashboardLayout from '@/components/layouts/DashboardLayout';
+import { CommonProvider } from '@/context/CommonContext';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
+      <CommonProvider>
         <DashboardLayout>
           {children}
         </DashboardLayout>
+      </CommonProvider>
   );
 }
